implemented sweetalert and styles in products image

+++ client/src/components/admin_page/admin_category_list/index.js
@@ -4,12 +4,31 @@ import { useDispatch } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { deleteCategory } from '../../../redux/actions/categories_actions';
 import { Btn, DataTable } from '../../styles/styled_global';
+import Swal from 'sweetalert2';
 
 const AdminCategoryList = ({ categories, language }) => {
 	const dispatch = useDispatch();
 	const handleDelete = (id) => {
-		dispatch(deleteCategory(id));
-		alert('Category deleted.');
+
+		Swal.fire({
+			heightAuto: false,
+			title: 'Confirm!',
+			text: 'Category deleted.',
+			icon: 'warning',
+			showCancelButton: true,
+			confirmButtonColor: '#3085d6',
+			cancelButtonColor: '#d33',
+			confirmButtonText: 'Yes, delete it!',
+		}).then((result) => {
+			if (result.isConfirmed) {
+				Swal.fire(
+					'Deleted!',
+					'Your file has been deleted.',
+					'success',
+					dispatch(deleteCategory(id))
+				)
+			}
+		})
 	}
 
 	return (
